Finished UI and Monsters
