+++ Assets/HealthBarUi.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 460f1e27735876f4aaf4ffc629296f47
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
