+++ Assets/HealthBarUi/HealthBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 119a2a5e107ba274c8540ac5592a3eeb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
