+++ Assets/Scripts/EnemyScripts/GhostDamage.cs
@@ -7,6 +7,7 @@ public class GhostDamage : MonoBehaviour
 {
     private Animator animator;
     public int health = 100;
+    public GameManager gameManager;
     private navigation_patrol patrolScript;
 
     void Start()
@@ -72,7 +73,7 @@ public class GhostDamage : MonoBehaviour
             rb.isKinematic = true; // Prevent further physics interactions
             rb.velocity = Vector3.zero; // Stop any residual movement
         }
-
+        gameManager.EnemyKilled();
         animator.SetBool("IsMoving", false);
     }
     private IEnumerator DisableSoundsAfterDelay()
